fix: make most adapters final and readonly
